[server] refactor server parts in separate packages
