+++ cmd/ingest-server/server/config/config.go
@@ -1,4 +1,4 @@
-package server
+package config
 
 import (
 	"encoding/json"
