+++ cmd/ingest-server/server/handlers/version_handler.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 )
 
-func (h Server) VersionHandler(w http.ResponseWriter, r *http.Request) {
+func VersionHandler(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	fmt.Fprintf(w, `{"version":"1.0.0"}`)
 }
